requirements.txt
torch>=2.0.0
transformers>=4.30.0
gradio>=3.50.0
PyGithub>=2.1.1
python-dotenv>=1.0.0
sentencepiece>=0.1.99
tokenizers>=0.13.3
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
black>=23.0.0
isort>=5.12.0
pytest>=7.4.0
mypy>=1.5.0
